Initialize node package and install jstodotxt and prompt-sync
The package was initialized using `yarn init -y`.

The jstodotxt package is used to parse todo items in the todo.txt so that
we can sort and filter them later. The prompt-sync package is used to
prompt the user for input in a simple and synchronous way. These packages
were installed using the command:

	yarn add jstodotxt prompt-sync